backend/requirements/base.txt
Django>=5.0
djangorestframework>=3.15
python-dotenv>=1.0
psycopg[binary]>=3.1
Pillow>=10.0
django-cors-headers>=4.3
whitenoise>=6.6
dj-database-url>=2.1
djangorestframework-simplejwt>=5.3
django-filter>=24.0
channels>=4.1
daphne>=4.1
sentry-sdk>=2.0
django-csp>=3.7
